refactor(openclaw-plugin): move setup.cli into setup/
