+++ packages/openclaw-plugin/src/setup/setup.cli.ts
